Clean up by moving different class to sub-modules and start to restrict outwards API
