+++ pyexasol/query_builders/csv/builders.py
@@ -1,15 +1,7 @@
-"""Build and validate CSV import and export SQL options.
-
-See the Exasol documentation for the `IMPORT <https://docs.exasol.com/db/latest/sql/import.htm>`_
-and `EXPORT <https://docs.exasol.com/db/latest/sql/export.htm>`_ statements.
-"""
-
 from __future__ import annotations
 
 from collections.abc import Iterable
-from dataclasses import dataclass
 from typing import (
-    TYPE_CHECKING,
     Annotated,
 )
 
@@ -20,41 +12,8 @@ from pydantic import (
     computed_field,
 )
 
-if TYPE_CHECKING:
-    from pyexasol import ExaFormatter
-
 ALLOWED_FORMAT = ("bz2", "csv", "gz", "zip")
 ALLOWED_TRIM = ("TRIM", "LTRIM", "RTRIM")
-
-
-@dataclass(frozen=True)
-class ClauseFormatter:
-    formatter: ExaFormatter
-
-    def column_delimiter(self, column_delimiter: str | None) -> str | None:
-        if column_delimiter is None:
-            return None
-        return f"COLUMN DELIMITER = {self.formatter.quote(column_delimiter)}"
-
-    def column_separator(self, column_separator: str | None) -> str | None:
-        if column_separator is None:
-            return None
-        return f"COLUMN SEPARATOR = {self.formatter.quote(column_separator)}"
-
-    def encoding(self, encoding: str | None) -> str | None:
-        if encoding is None:
-            return None
-        return f"ENCODING = {self.formatter.quote(encoding)}"
-
-    def null(self, null: str | None) -> str | None:
-        if null is None:
-            return None
-        return f"NULL = {self.formatter.quote(null)}"
-
-    def row_separator(self, row_separator: str | None) -> str | None:
-        if row_separator is None:
-            return None
-        return f"ROW SEPARATOR = {self.formatter.quote(row_separator)}"
 
 
 def validate_format(file_format: str | None) -> str | None:
